+++ s/signals/index.ts
@@ -6,6 +6,9 @@ export * from "./core/effect.js"
 export * from "./core/lazy.js"
 export * from "./core/signal.js"
 
+export * from "./rz/rz-map.js"
+export * from "./rz/rz-set.js"
+
 export * from "./porcelain.js"
 export * from "./types.js"